fix pr feedback
