+++ src/renderer/utils/localstorage-clear.tsx
@@ -1,20 +1,28 @@
 import { useTranslation } from 'react-i18next';
 import { useEffectOnce } from 'react-use';
 
-import { RETAINED_LOCALSTORAGE_KEYS } from '../../common/env-and-consts';
+export const RETAINED_LOCALSTORAGE_KEYS = [
+  'AieApp$keyOctave',
+  'AieApp$projectFolder',
+  'AieApp$recordingProject',
+  'AieApp$recordingSet',
+  'AieApp$recordingList',
+  'ConfigureRecordingSetPage$recordingSets',
+  'ConfigureRecordingSetPage$projectFile',
+  'RecordingPage$index',
+];
 
-const useResumeCheck = (func: () => void) => {
+export const useResumeCheck = (goToRecordingPage: () => void) => {
   const { t } = useTranslation();
   useEffectOnce(() => {
     const reservedStateValues = RETAINED_LOCALSTORAGE_KEYS.map((state) => localStorage.getItem(state));
     if (!reservedStateValues.includes(null)) {
       if (confirm(`${t('Do you want to continue from where you left off?')}`)) {
-        func();
+        goToRecordingPage();
       } else {
         RETAINED_LOCALSTORAGE_KEYS.forEach((e) => localStorage.removeItem(e));
       }
     }
   });
+  return null;
 };
-
-export default useResumeCheck;
